add more tests

+++ internal/handlers/common/filter.go
@@ -340,6 +340,7 @@ func filterOperator(doc *types.Document, operator string, filterValue any) (bool
 // If result of processing `$expr` is null, zero value number or boolean false,
 // it returns false indicating filter not matched. For other values, it returns true.
 func filterExprOperator(doc, filter *types.Document) (bool, error) {
+	// TODO https://github.com/FerretDB/FerretDB/issues/3170
 	op, err := operators.NewExpr(filter, "$expr")
 	if err != nil {
 		return false, err
@@ -354,7 +355,7 @@ func filterExprOperator(doc, filter *types.Document) (bool, error) {
 	case *types.Document, *types.Array, string, types.Binary, types.ObjectID, time.Time, types.Regex, types.Timestamp:
 		return true, nil
 	case float64, int32, int64:
-		if res := types.Compare(v, float64(0)); res == types.Equal {
+		if res := types.Compare(v, int32(0)); res == types.Equal {
 			return false, nil
 		}
